requirements.txt
alembic==1.13.2
aniso8601==9.0.1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dnspython==2.6.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-PyMongo==2.3.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
gunicorn==23.0.0
holidays==0.56
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.4
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
plotly==5.24.0
pymongo==4.8.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rpds-py==0.20.0
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.32
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.4
xgboost==2.1.1
zipp==3.20.1
